+++ docs/CHANGELOG.md
@@ -33,8 +33,10 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `SystemDesc` proc macro supports `#[system_desc(event_reader_id)]` to register event reader. ([#1883])
 * `SystemDesc` proc macro supports `#[system_desc(flagged_storage_reader(Component))]`. ([#1886])
 * `DispatcherOperation` stores system name and dependencies as `String`s. ([#1891])
-* `TextureProcessor` renamed to `TextureProcessorSystem` ([#1839]).
-* `MeshProcessor` renamed to `MeshProcessorSystem` ([#1839]).
+* `TextureProcessor` renamed to `TextureProcessorSystem`. ([#1839])
+* `MeshProcessor` renamed to `MeshProcessorSystem`. ([#1839])
+* `AmethystApplication::with_setup` now takes in `FnOnce(&mut World) + Send + 'static`. ([#1912])
+* `AmethystApplication::with_setup` runs the function before the dispatcher. ([#1912])
 
 ### Fixed
 
@@ -51,6 +53,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1891]: https://github.com/amethyst/amethyst/pull/1891
 [#1896]: https://github.com/amethyst/amethyst/pull/1896
 [#1839]: https://github.com/amethyst/amethyst/pull/1839
+[#1912]: https://github.com/amethyst/amethyst/pull/1912
 
 ## [0.12.0] - 2019-07-30